Added New Winning Game Logic
Received guidance from Jordan to help build out indicator for winning
the game. Created a new gameBoard array which logs the X's and O's. It
then checks on the position of the array and whether or not it contains
an X. Will need to further build this out for Columns and Diagonals, as
well as for Player O wins.